+++ app/controllers/WeekController.ts
@@ -40,6 +40,10 @@ export class WeekController implements ng.IController {
         window.addEventListener("beforeunload", () => {
             this.$interval.cancel(this.loadInterval); // Cancel the load interval on close
         });
+
+        this.$scope.$on("logout", () => {
+            this.$interval.cancel(this.loadInterval);
+        });
     }
 
     /** Comparator to compare two events inside a day */
@@ -59,10 +63,10 @@ export class WeekController implements ng.IController {
         this.hounds.load(this.week.getDay(0));
     }
 
-    /** 
+    /**
      * Get the text from an event to display inside a day
      * @param event Schedule event to get text
-     * 
+     *
      * @returns string to put in day
      */
     public getEventText(event: IScheduleEvent): string {
@@ -70,11 +74,20 @@ export class WeekController implements ng.IController {
             return event.text;
         }
 
+        const startHour = event.startDate.getHours();
+        const openingAm = this.$settings.HOURS.OPENING.AM;
+        const closingAm = this.$settings.HOURS.CLOSING.AM;
+        const openingPm = this.$settings.HOURS.OPENING.PM;
+        const closingPm = this.$settings.HOURS.CLOSING.PM;
+
         // When there is no end time information set it to the closing hours
         if (dates.compareAsc(event.startDate, event.endDate) === 0) {
             if (event.type === DEFAULT.CONSTANTS.DAYCARE) {
-                event.endDate.setHours(this.$settings.HOURS.CLOSING.PM);
-            } else {
+                event.endDate.setHours(closingPm);
+            } else if (
+                this.isInInterval(startHour, openingAm, closingAm) ||
+                this.isInInterval(startHour, openingPm, closingPm)
+            ) {
                 event.endDate = new Date(event.startDate);
                 const closingHour =
                     event.startDate.getHours() < 12
@@ -86,5 +99,23 @@ export class WeekController implements ng.IController {
 
         // Call hounds library to retrieve time prepend
         return getTimePrepend(event) + " " + event.text;
+    }
+
+    /**
+     * Returns whether or not x is in interval [a, b] inclusive
+     * @param x number to check
+     * @param a lower bound
+     * @param b upper bound
+     */
+    private isInInterval(x: number, a: number, b: number): boolean {
+        if (a === b) {
+            return false;
+        } else if (a > b) {
+            const temp = b;
+            b = a;
+            a = temp;
+        }
+
+        return x >= a && x <= b;
     }
 }