+++ changes.md
@@ -17,4 +17,10 @@
 - Drop down menu for editing bookings
 - Move Grooming into dog profile
 
+- Add an event "copy" function
+
+- Update to new angular
+
+- Check API Responses and alert user on an error
+
 ## ERRORS